Insert query added.

+++ src/UserPrompt.java
@@ -1,26 +1,24 @@
+package QueryParser;
+
 import Model.IQuery;
 import Model.Result;
-import QueryParser.DatabaseHelper;
 import common.CatalogDB;
 import common.Constants;
-import storage.StorageManager;
 
 import java.io.File;
 import java.util.Scanner;
 
 public class UserPrompt {
 
+    public static final String USE_HELP_MESSAGE = "Please use 'HELP' to see a list of commands";
     static boolean isExit = false;
     static Scanner scanner = new Scanner(System.in).useDelimiter(";");
 
     public static void main(String[] args) {
 
+        initializeDatabase();
 		splashScreen();
         String userCommand = "";
-
-        // Initialize the database.
-        initializeDatabase();
-        //new Test().run(1);
 
         while(!isExit) {
             System.out.print(DatabaseHelper.prompt);
@@ -43,8 +41,13 @@ public class UserPrompt {
 		
 		if(userCommand.toLowerCase().equals(DatabaseHelper.SHOW_TABLES_COMMAND.toLowerCase())){
 		    IQuery query = DatabaseHelper.ShowTableListQueryHandler();
-		    ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
 		    return;
+        }
+        else if(userCommand.toLowerCase().equals(DatabaseHelper.SHOW_DATABASES_COMMAND.toLowerCase())){
+            IQuery query = DatabaseHelper.ShowDatabaseListQueryHandler();
+            DatabaseHelper.ExecuteQuery(query);
+            return;
         }
         else if(userCommand.toLowerCase().equals(DatabaseHelper.HELP_COMMAND.toLowerCase())){
             DatabaseHelper.HelpQueryHandler();
@@ -56,17 +59,50 @@ public class UserPrompt {
         }
         else if(userCommand.toLowerCase().equals(DatabaseHelper.EXIT_COMMAND.toLowerCase()) ||
                 userCommand.toLowerCase().equals(DatabaseHelper.QUIT_COMMAND.toLowerCase())){
+
             System.out.println("Exiting Database...");
             isExit = true;
             return;
         }
-        else if(userCommand.toLowerCase().startsWith(DatabaseHelper.DROP_COMMAND.toLowerCase())){
-            String tableName = userCommand.substring(DatabaseHelper.DROP_COMMAND.length());
+        else if(userCommand.toLowerCase().startsWith(DatabaseHelper.USE_DATABASE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.USE_DATABASE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
+            String databaseName = userCommand.substring(DatabaseHelper.USE_DATABASE_COMMAND.length());
+            IQuery query = DatabaseHelper.UseDatabaseQueryHandler(databaseName.trim());
+            DatabaseHelper.ExecuteQuery(query);
+            return;
+        }
+        else if(userCommand.toLowerCase().startsWith(DatabaseHelper.DROP_TABLE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.DROP_TABLE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
+            String tableName = userCommand.substring(DatabaseHelper.DROP_TABLE_COMMAND.length());
             IQuery query = DatabaseHelper.DropTableQueryHandler(tableName.trim());
-            ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
+            return;
+        }
+        else if(userCommand.toLowerCase().startsWith(DatabaseHelper.DROP_DATABASE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.DROP_DATABASE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
+            String databaseName = userCommand.substring(DatabaseHelper.DROP_DATABASE_COMMAND.length());
+            IQuery query = DatabaseHelper.DropDatabaseQueryHandler(databaseName.trim());
+            DatabaseHelper.ExecuteQuery(query);
             return;
         }
         else if(userCommand.toLowerCase().startsWith(DatabaseHelper.SELECT_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.SELECT_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
             int index = userCommand.toLowerCase().indexOf("from");
             if(index == -1) {
                 DatabaseHelper.UnrecognisedCommand(userCommand, "Expected FROM keyword");
@@ -80,16 +116,21 @@ public class UserPrompt {
             if(index == -1) {
                 String tableName = restUserQuery.trim();
                 IQuery query = DatabaseHelper.SelectQueryHandler(attributeList.split(","), tableName, "");
-                ExecuteQuery(query);
+                DatabaseHelper.ExecuteQuery(query);
                 return;
             }
 
             String tableName = restUserQuery.substring(0, index);
             String conditions = restUserQuery.substring(index + "where".length());
             IQuery query = DatabaseHelper.SelectQueryHandler(attributeList.split(","), tableName.trim(), conditions);
-            ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
         }
         else if(userCommand.toLowerCase().startsWith(DatabaseHelper.INSERT_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.INSERT_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
             String tableName = "";
             String columns = "";
 
@@ -130,16 +171,21 @@ public class UserPrompt {
 
             valuesList = valuesList.substring(1, valuesList.length()-1);
             IQuery query = DatabaseHelper.InsertQueryHandler(tableName, columns, valuesList);
-            ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
         }
         else if(userCommand.toLowerCase().startsWith(DatabaseHelper.DELETE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.DELETE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
             String tableName = "";
             String condition = "";
             int index = userCommand.toLowerCase().indexOf("where");
             if(index == -1) {
                 tableName = userCommand.substring(DatabaseHelper.DELETE_COMMAND.length()).trim();
                 IQuery query = DatabaseHelper.DeleteQueryHandler(tableName, condition);
-                ExecuteQuery(query);
+                DatabaseHelper.ExecuteQuery(query);
                 return;
             }
 
@@ -149,9 +195,14 @@ public class UserPrompt {
 
             condition = userCommand.substring(index + "where".length());
             IQuery query = DatabaseHelper.DeleteQueryHandler(tableName, condition);
-            ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
         }
         else if(userCommand.toLowerCase().startsWith(DatabaseHelper.UPDATE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.UPDATE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
             String tableName = "";
             String conditions = "";
             int setIndex = userCommand.toLowerCase().indexOf("set");
@@ -165,16 +216,32 @@ public class UserPrompt {
             int whereIndex = userCommand.toLowerCase().indexOf("where");
             if(whereIndex == -1){
                 IQuery query = DatabaseHelper.UpdateQuery(tableName, clauses, conditions);
-                ExecuteQuery(query);
+                DatabaseHelper.ExecuteQuery(query);
                 return;
             }
 
             clauses = userCommand.substring(setIndex + "set".length(), whereIndex).trim();
             conditions = userCommand.substring(whereIndex + "where".length());
             IQuery query = DatabaseHelper.UpdateQuery(tableName, clauses, conditions);
-            ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
+        }
+        else if(userCommand.toLowerCase().startsWith(DatabaseHelper.CREATE_DATABASE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.CREATE_DATABASE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
+            String databaseName = userCommand.substring(DatabaseHelper.CREATE_DATABASE_COMMAND.length());
+            IQuery query = DatabaseHelper.CreateDatabaseQueryHandler(databaseName.trim());
+            DatabaseHelper.ExecuteQuery(query);
+            return;
         }
         else if(userCommand.toLowerCase().startsWith(DatabaseHelper.CREATE_TABLE_COMMAND.toLowerCase())){
+            if(!PartsEqual(userCommand, DatabaseHelper.CREATE_TABLE_COMMAND)){
+                DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
+                return;
+            }
+
             String tableName = "";
             IQuery query = null;
 
@@ -193,22 +260,27 @@ public class UserPrompt {
             String columnsPart = userCommand.substring(openBracketIndex + 1, userCommand.length()-1);
 
             query = DatabaseHelper.CreateTableQueryHandler(tableName, columnsPart);
-            ExecuteQuery(query);
+            DatabaseHelper.ExecuteQuery(query);
             return;
         }
         else{
-            DatabaseHelper.UnrecognisedCommand(userCommand, "Please use 'HELP' to see a list of commands");
+            DatabaseHelper.UnrecognisedCommand(userCommand, USE_HELP_MESSAGE);
         }
     }
 
-    private static void ExecuteQuery(IQuery query) {
-        if(query!= null && query.ValidateQuery()){
-            Result result = query.ExecuteQuery();
-            result.Display();
+    private static boolean PartsEqual(String userCommand, String expectedCommand) {
+        String[] userParts = userCommand.toLowerCase().split(" ");
+        String[] actualParts = expectedCommand.toLowerCase().split(" ");
+
+        for(int i=0;i<actualParts.length;i++){
+            if(!actualParts[i].equals(userParts[i])){
+                return false;
+            }
         }
+
+        return true;
     }
 
-    // Initialize the database.
     public static void initializeDatabase() {
         File baseDir = new File(Constants.DEFAULT_DATA_DIRNAME);
         if(baseDir.exists()) {
@@ -217,23 +289,6 @@ public class UserPrompt {
                 if(catalogDir.mkdir()) {
                     new CatalogDB().createCatalogDB();
                 }
-            }
-        }
-
-        // Check if the default database exists.
-        if (StorageManager.defaultDatabaseExists()) {
-            // Default database exist.
-            //System.out.println("Database already created.");
-
-        }
-        else  {
-            StorageManager manager = new StorageManager();
-            boolean create = manager.createDatabase(Constants.DEFAULT_USER_DATABASE);
-            if (create) {
-                //System.out.println("Created new database.");
-            }
-            else {
-                //System.out.println("Failed to create new database.");
             }
         }
